update basic function ordering
