+++ ndsl/stencils/__init__.py
@@ -1,31 +1,27 @@
 from .basic_operations import (
     add,
-    add_2d,
     add_to_self,
-    add_to_self_2d,
-    adjust_divide_stencil,
-    adjustmentfactor_stencil,
-    average_in,
     copy,
-    copy_2d,
-    dim,
     divide,
-    divide_2d,
-    divide_to_self,
-    divide_to_self_2d,
+    divide_self,
     multiply,
-    multiply_2d,
     multiply_to_self,
-    multiply_to_self_2d,
     select_k,
-    set_IJ_mask_value,
     set_value,
-    set_value_2D,
-    sign,
     subtract,
+    subtract_from_self,
+)
+from .basic_operations_2d import (
+    add_2d,
+    add_to_self_2d,
+    copy_2d,
+    divide_2d,
+    divide_self_2d,
+    multiply_2d,
+    multiply_to_self_2d,
+    set_value_2d,
     subtract_2d,
-    subtract_to_self,
-    subtract_to_self_2d,
+    subtract_from_self_2d,
 )
 from .corners import FillCornersBGrid
 
@@ -33,30 +29,24 @@ from .corners import FillCornersBGrid
 __all__ = [
     "FillCornersBGrid",
     "add",
-    "add_2d",
     "add_to_self",
-    "add_to_self_2d",
-    "adjust_divide_stencil",
-    "adjustmentfactor_stencil",
-    "average_in",
     "copy",
-    "copy_2d",
-    "dim",
     "divide",
-    "divide_2d",
-    "divide_to_self",
-    "divide_to_self_2d",
+    "divide_self",
     "multiply",
-    "multiply_2d",
     "multiply_to_self",
-    "multiply_to_self_2d",
     "select_k",
-    "set_IJ_mask_value",
     "set_value",
-    "set_value_2D",
-    "sign",
     "subtract",
+    "subtract_from_self",
+    "add_2d",
+    "add_to_self_2d",
+    "copy_2d",
+    "divide_2d",
+    "divide_self_2d",
+    "multiply_2d",
+    "multiply_to_self_2d",
+    "set_value_2d",
     "subtract_2d",
-    "subtract_to_self",
-    "subtract_to_self_2d",
+    "subtract_from_self_2d",
 ]
