+++ ndsl/stencils/arithmetical_functions.py
@@ -1,0 +1,27 @@
+import typing
+
+from ndsl.dsl.gt4py import function
+
+
+@typing.no_type_check
+@function
+def sign(a, b):
+    """
+    Defines a_sign_b as the absolute value of a, and checks if b is positive or
+    negative, assigning the analogous sign value to a_sign_b. a_sign_b is returned.
+
+    Args:
+        a: A number
+        b: A number
+    """
+    a_sign_b = abs(a)
+    return a_sign_b if b > 0 else -a_sign_b
+
+
+@typing.no_type_check
+@function
+def dim(a, b):
+    """
+    Calculates a - b, camped to 0, i.e. max(a - b, 0).
+    """
+    return max(a - b, 0)
