+++ ndsl/stencils/basic_operations.py
@@ -1,13 +1,8 @@
-import typing
-
-from ndsl.dsl.gt4py import FORWARD, PARALLEL, computation, function, interval
+from ndsl.dsl.gt4py import FORWARD, K, PARALLEL, computation, interval
 from ndsl.dsl.typing import (
-    Bool,
-    BoolFieldIJ,
     Float,
     FloatField,
     FloatFieldIJ,
-    IntField,
     IntFieldIJ,
 )
 
@@ -21,18 +16,6 @@ def copy(input: FloatField, output: FloatField) -> None:
         output: output field
     """
     with computation(PARALLEL), interval(...):
-        output = input
-
-
-def copy_2d(input: FloatFieldIJ, output: FloatFieldIJ) -> None:
-    """
-    Copy one field into another - 2D variant.
-
-    Args:
-        input: input field
-        output: output field
-    """
-    with computation(FORWARD), interval(...):
         output = input
 
 
@@ -61,31 +44,6 @@ def add_to_self(field: FloatField, summand: FloatField) -> None:
         field = field + summand
 
 
-def add_2d(summand_1: FloatFieldIJ, summand_2: FloatFieldIJ, sum: FloatFieldIJ) -> None:
-    """
-    Add two inputs together, sum to a new field - 2D variant.
-
-    Args:
-        summand_1: input field
-        summand_2: input field
-        sum: output field
-    """
-    with computation(FORWARD), interval(0, 1):
-        sum = summand_1 + summand_2
-
-
-def add_to_self_2d(field: FloatFieldIJ, summand: FloatFieldIJ) -> None:
-    """
-    Add a summand to a field - 2D variant.
-
-    Args:
-        field: field to be modifid
-        summand: modification to be made
-    """
-    with computation(FORWARD), interval(0, 1):
-        field = field + summand
-
-
 def subtract(
     minuend: FloatField, subtrahend: FloatField, difference: FloatField
 ) -> None:
@@ -101,7 +59,7 @@ def subtract(
         difference = minuend - subtrahend
 
 
-def subtract_to_self(field: FloatField, subtrahend: FloatField) -> None:
+def subtract_from_self(field: FloatField, subtrahend: FloatField) -> None:
     """
     Subtract a subtrahend from a field.
 
@@ -110,33 +68,6 @@ def subtract_to_self(field: FloatField, subtrahend: FloatField) -> None:
         subtrahend: modification to be made
     """
     with computation(PARALLEL), interval(...):
-        field = field - subtrahend
-
-
-def subtract_2d(
-    minuend: FloatFieldIJ, subtrahend: FloatFieldIJ, difference: FloatFieldIJ
-) -> None:
-    """
-    Subtract summand_2 from minuend, output to a new field - 2D variant.
-
-    Args:
-        minuend: input field
-        summand_2: input field
-        difference: output field
-    """
-    with computation(FORWARD), interval(0, 1):
-        difference = minuend - subtrahend
-
-
-def subtract_to_self_2d(field: FloatFieldIJ, subtrahend: FloatFieldIJ) -> None:
-    """
-    Subtract a modification from a field - 2D variant.
-
-    Args:
-        field: field to be modifid
-        subtrahend: modification to be made
-    """
-    with computation(FORWARD), interval(0, 1):
         field = field - subtrahend
 
 
@@ -165,33 +96,6 @@ def multiply_to_self(field: FloatField, factor: FloatField) -> None:
         field = field * factor
 
 
-def multiply_2d(
-    factor_1: FloatFieldIJ, factor_2: FloatFieldIJ, product: FloatFieldIJ
-) -> None:
-    """
-    Multiply two inputs together, output to a new field - 2D variant.
-
-    Args:
-        factor_1: input field
-        factor_2: input field
-        output: output field
-    """
-    with computation(FORWARD), interval(0, 1):
-        product = factor_1 * factor_2
-
-
-def multiply_to_self_2d(field: FloatFieldIJ, factor: FloatFieldIJ) -> None:
-    """
-    Muultiply a field by a factor - 2D variant.
-
-    Args:
-        field: field to be modifid
-        factor: modification factor
-    """
-    with computation(FORWARD), interval(0, 1):
-        field = field * factor
-
-
 def divide(dividend: FloatField, divisor: FloatField, quotient: FloatField) -> None:
     """
     Divide dividend by divisor, output to a new field.
@@ -205,7 +109,7 @@ def divide(dividend: FloatField, divisor: FloatField, quotient: FloatField) -> N
         quotient = dividend / divisor
 
 
-def divide_to_self(field: FloatField, divisor: FloatField) -> None:
+def divide_self(field: FloatField, divisor: FloatField) -> None:
     """
     Muultiply a field by a factor - 2D variant.
 
@@ -214,33 +118,6 @@ def divide_to_self(field: FloatField, divisor: FloatField) -> None:
         divisor: modification factor
     """
     with computation(PARALLEL), interval(...):
-        field = field / divisor
-
-
-def divide_2d(
-    dividend: FloatFieldIJ, divisor: FloatFieldIJ, quotient: FloatFieldIJ
-) -> None:
-    """
-    Divide dividend by divisor, output to a new field - 2D variant.
-
-    Args:
-        dividend: input field
-        divisor: input field
-        quotient: output field
-    """
-    with computation(FORWARD), interval(0, 1):
-        quotient = dividend / divisor
-
-
-def divide_to_self_2d(field: FloatFieldIJ, divisor: FloatFieldIJ) -> None:
-    """
-    Muultiply a field by a factor - 2D variant.
-
-    Args:
-        field: field to be modifid
-        divisor: modification factor
-    """
-    with computation(FORWARD), interval(0, 1):
         field = field / divisor
 
 
@@ -256,58 +133,9 @@ def set_value(field: FloatField, value: Float) -> None:
         field = value
 
 
-def set_value_2D(field: FloatFieldIJ, value: Float) -> None:
-    """
-    Sets every element of a field to a single value - 2D variant.
-
-    Args:
-        field: output field
-        value: value of Float type
-    """
-    with computation(FORWARD), interval(...):
-        field = value
-
-
-def set_IJ_mask_value(field: BoolFieldIJ, value: Bool) -> None:
-    """
-    Sets every element of buffer to either True or False.
-
-    Args:
-        field: output field
-        value: value of Bool type
-    """
-    with computation(FORWARD), interval(0, 1):
-        field = value
-
-
-def adjustmentfactor_stencil(adjustment: FloatFieldIJ, field: FloatField) -> None:
-    """
-    Multiplies a field by an adjustment factor, modifying the original field.
-
-    Args:
-        adjustment: adjustment factor
-        field: field to be modified
-    """
-    with computation(PARALLEL), interval(...):
-        field = field * adjustment
-
-
-def adjust_divide_stencil(adjustment: FloatField, field: FloatField) -> None:
-    """
-    Divides a field by an adjustment factor, modifying the original field.
-
-    Args:
-        adjustment: adjustment factor
-        field: field to be modified
-    """
-    with computation(PARALLEL), interval(...):
-        field = field / adjustment
-
-
 def select_k(
     in_field: FloatField,
     out_field: FloatFieldIJ,
-    k_mask: IntField,
     k_select: IntFieldIJ,
 ) -> None:
     """
@@ -317,17 +145,15 @@ def select_k(
     Args:
         in_field: A 3D array to select from
         out_field: A 2D field to save values in
-        k_mask: a field that lists each k-index
         k_select: the k-value to extract from in_field
     """
-    # TODO: refactor this using THIS_K instead of a mask
     with computation(FORWARD), interval(...):
-        if k_mask == k_select:
+        if K == k_select:
             out_field = in_field
 
 
-def average_in(
-    q_out: FloatField,
+def average_input(
+    field: FloatField,
     adjustment: FloatField,
 ) -> None:
     """
@@ -339,28 +165,4 @@ def average_in(
         q_out: output field
     """
     with computation(PARALLEL), interval(...):
-        q_out = (q_out + adjustment) * 0.5
-
-
-@typing.no_type_check
-@function
-def sign(a, b):
-    """
-    Defines a_sign_b as the absolute value of a, and checks if b is positive or
-    negative, assigning the analogous sign value to a_sign_b. a_sign_b is returned.
-
-    Args:
-        a: A number
-        b: A number
-    """
-    a_sign_b = abs(a)
-    return a_sign_b if b > 0 else -a_sign_b
-
-
-@typing.no_type_check
-@function
-def dim(a, b):
-    """
-    Calculates a - b, camped to 0, i.e. max(a - b, 0).
-    """
-    return max(a - b, 0)
+        field = (field + adjustment) * 0.5
